release

+++ src/vm/sequencers/native/op_free_get.go
@@ -33,7 +33,7 @@ func NewOpFreeGet(vm core.IVM, op *bytecode.Opcodes) (core.IOpExecutor, error) {
 func (op *OpFreeGet) Execute(decoder *core.Decoder) {
 	// Operands Offset 1 (8-bit)
 	freeIndex := decoder.Read(0)
-	freeVar := op.vm.Frame().FreeVarsIndex(freeIndex)
+	freeVar := op.vm.Frame().FreeVarsIndex(uint(freeIndex))
 	if freeVar == nil {
 		op.vm.SetError(fmt.Errorf("free variable %d not found", freeIndex))
 		return
